Bugfix #279 	Bad return value for ldns_udp_connect and ldns_tcp_connect
ldns_udp_connect, ldns_tcp_connect, ldns_udp_bgsend and ldns_tcp_bgsend return
a socket and otherwise 0 on failure.  This is wrong because 0 can be a valid
socket.

To not mess with the existing API, ldns_udp_connect2, ldns_tcp_connect2,
ldns_udp_bgsend2 and ldns_tcp_bgsend2 are introduced that do the same but
return -1 on failure.  Internal usage of the old functions replaced with the
new ones.